+++ NEWS.md
@@ -1,3 +1,5 @@
+# ncdfCF (development version)
+
 # ncdfCF 0.8.0
 
 #### Conventions