+++ NEWS.md
@@ -1,4 +1,4 @@
-# ncdfCF (development version)
+# ncdfCF 0.8.1
 
 Patch release with bug fixes and some new functionality.